Move the donations route logic to a route folder
